Reorganized TemplateTrackSettings to match framework folder structure
