+++ frontend/src/modules/WellLogViewer/settings/components/TemplateTrackSettings/index.ts
@@ -1,0 +1,1 @@
+export { TemplateTrackSettings } from "./templateTrackSettings";
